able to read vertices

+++ src/core.nim
@@ -11,6 +11,6 @@ proc `+`*(lhs, rhs: Point) : Point =
   result.x = lhs.x + rhs.x
   result.y = lhs.y + rhs.y
 
-proc `*`*(lhs : float64, rhs: Point) : Point = 
+proc `*`*(lhs : float32, rhs: Point) : Point = 
   result.x = (lhs * rhs.x.toFloat).toInt
   result.y = (lhs * rhs.y.toFloat).toInt
